requirements.txt
tensorflow==1.6.0
numpy==1.14.1
Pillow==5.0.0
Django==2.0.3
